Add tests for bitmap composition and stroke rendering
